sbe_chipOp_handler: Improve read behaviour
Prepare to read until we have read all data, and in the mean time, ensure
we've prepared a read buffer at least as long as the expected response.

Note that the actual response may also be either smaller or greater in
size than the expected response[1]. Currently there's no proper way to
handle these cases as the SBEFIFO driver issues EAGAIN to a read() when
all data has been read instead of returning 0. As a consequence we have
no correct way to extract all the data from the FIFO, as the only means
to tell we've read all the data is for a short read to occur. Given the
read syscall interface allows for short reads to occur normally (i.e. a
short read does not imply all data has been read) we reach an ambiguous
state. We can't rely on EAGAIN for similar reasons.

At the moment, we assume all data has been read on a short read.

Further, improve the error handling by adding description tables for the
primary and secondary status error codes. This improves the output of
applications using the library.

[1] If the request is not correctly formulated then an error status will
be returned with no data preceding the metadata. If an error occurred
whilst the request was being service then we may receive FFDC data in
addition to the data and metadata already provided.

Signed-off-by: Andrew Jeffery <andrew@aj.id.au>